+++ extension-opera/opena11y-for-ainspector.js
@@ -150,7 +150,7 @@
   /* Constants */
   const debug$18 = new DebugLogging('constants', false);
 
-  const VERSION = '2.0.6';
+  const VERSION = '2.0.7';
 
   /**
    * @constant RULESET
@@ -40398,6 +40398,25 @@
   debug.flag = false;
 
   /**
+   * @function aiAllRuleResults
+   *
+   * @desc Returns an array of information containing the rule results for the all the rules
+   * *
+   * @return {Array}  see description
+   */
+
+  function aiAllRuleResults (allRuleResults) {
+    const rule_results = [];
+
+    allRuleResults.forEach( rr => {
+      const result = getRuleResultInfo(rr);
+      rule_results.push(result);
+    });
+    return rule_results;
+  }
+
+
+  /**
    * @function aiRuleResultsByCategory
    *
    * @desc Returns an array of information containing the rule results for the rules
@@ -40473,14 +40492,42 @@
    */
 
   function getRuleResultInfo (rule_result) {
-    return {      id: rule_result.rule.getId(),
-              id_nls: rule_result.rule.getIdNLS(),
-             summary: rule_result.rule.getSummary(),
-              result: rule_result.getResultValueNLS(),
-        result_value: rule_result.getResultValue(),
-                  sc: rule_result.rule.getPrimarySuccessCriterionId(),
-               level: rule_result.rule.getWCAGLevel(),
-            required: rule_result.isRuleRequired()
+
+    const summary = rule_result.getResultsSummary();
+
+    return {                id: rule_result.rule.getId(),
+                        id_nls: rule_result.rule.getIdNLS(),
+                  rule_summary: rule_result.rule.getSummary(),
+
+              result_value_nls: rule_result.getResultValueNLS(),
+                  result_value: rule_result.getResultValue(),
+                result_message: rule_result.getResultMessage(),
+
+                 guideline_nls: rule_result.rule.getGuidelineInfo().title,
+                guideline_code: rule_result.rule.getGuidelineInfo().id,
+
+          success_criteria_nls: rule_result.rule.getPrimarySuccessCriterionInfo().title,
+         success_criteria_code: rule_result.rule.getPrimarySuccessCriterionInfo().id,
+             rule_category_nls: rule_result.rule.getCategoryInfo().title,
+                    wcag_level: rule_result.rule.getWCAGLevel(),
+
+                 rule_required: rule_result.isRuleRequired(),
+
+           rule_scope_code_nls: rule_result.rule.getScopeNLS(),
+               rule_scope_code: rule_result.rule.getScope(),
+
+          implementation_score: rule_result.getImplementationScore(),
+          implementation_value: rule_result.getImplementationValue(),
+            implementation_nls: rule_result.getImplementationValueNLS(),
+
+             results_violation: summary.violations,
+               results_warning: summary.warnings,
+               results_failure: summary.warnings + summary.violations,
+          results_manual_check: summary.manual_checks,
+                results_passed: summary.passed,
+                results_hidden: summary.hidden,
+
+                    has_hidden: rule_result.hasHiddenElementResults()
         };
   }
 
@@ -40703,6 +40750,7 @@
             response.rule_summary          = er.ruleResultSummary.data;
             response.rc_rule_results_group = er.rcRuleGroupResults.data;
             response.gl_rule_results_group = er.glRuleGroupResults.data;
+            response.rule_results          = aiAllRuleResults(er.allRuleResults);
             break;
 
           case 'rule-group':